requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
